pyproject.toml
[tool.poetry]
name = "render_demo"
version = "0.1.0"
description = "Test app deployment on render.com"
authors = ["le_woudar <lewoudar@gmail.com>"]
license = "MIT"

packages = [
    { include = "render_demo" }
]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.85.0"
tortoise-orm = "^0.19.2"
hypercorn = "^0.14.3"

[tool.poetry.dev-dependencies]
httpie = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
